Silence compiler warnings (10/11).
svn path=/trunk/; revision=37832

+++ reactos/boot/freeldr/freeldr/fs/ntfs.c
@@ -258,7 +258,7 @@ static ULONGLONG NtfsReadAttribute(PNTFS_ATTR_CONTEXT Context, ULONGLONG Offset,
         {
             CurrentOffset += DataRunLength * NtfsClusterSize;
             DataRun = NtfsDecodeRun(DataRun, &DataRunOffset, &DataRunLength);
-            if (DataRunLength != -1)
+            if (DataRunLength != (ULONGLONG)-1)
             {
                 DataRunStartLCN = LastLCN + DataRunOffset;
                 LastLCN = DataRunStartLCN;
